# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AlertCommentRequest" : {
        "properties" : {
          "comment" : {
            "description" : "Comentario para el historial de la alerta",
            "example" : "Revisando estados de cuenta",
            "maxLength" : 2000,
            "minLength" : 0,
            "pattern" : "\\s*|[\\p{L}\\p{N}\\r\\n\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*[\\p{L}\\p{N}][\\p{L}\\p{N}\\r\\n\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AlertDetailResponse" : {
        "properties" : {
          "accountId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "assignedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "assigneeId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "assigneeName" : {
            "type" : "string"
          },
          "closedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "criticality" : {
            "enum" : [ "CRITICAL", "HIGH", "MEDIUM", "LOW" ],
            "type" : "string"
          },
          "evidence" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "maskedAccount" : {
            "type" : "string"
          },
          "raisedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "resolution" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "UNASSIGNED", "ASSIGNED", "IN_REVIEW", "CLOSED" ],
            "type" : "string"
          },
          "typeCode" : {
            "type" : "string"
          },
          "typeDescription" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AlertHistoryResponse" : {
        "properties" : {
          "actorName" : {
            "type" : "string"
          },
          "assigneeName" : {
            "type" : "string"
          },
          "comment" : {
            "type" : "string"
          },
          "newStatus" : {
            "enum" : [ "UNASSIGNED", "ASSIGNED", "IN_REVIEW", "CLOSED" ],
            "type" : "string"
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "previousStatus" : {
            "enum" : [ "UNASSIGNED", "ASSIGNED", "IN_REVIEW", "CLOSED" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AlertResponse" : {
        "properties" : {
          "account" : {
            "type" : "string"
          },
          "assigneeId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "assigneeName" : {
            "type" : "string"
          },
          "criticality" : {
            "enum" : [ "CRITICAL", "HIGH", "MEDIUM", "LOW" ],
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "raisedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "UNASSIGNED", "ASSIGNED", "IN_REVIEW", "CLOSED" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AlertTypeResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "defaultCriticality" : {
            "enum" : [ "CRITICAL", "HIGH", "MEDIUM", "LOW" ],
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CloseAlertRequest" : {
        "properties" : {
          "comment" : {
            "description" : "Comentario para el historial de la alerta",
            "example" : "Pago de aguinaldo",
            "maxLength" : 2000,
            "minLength" : 0,
            "pattern" : "\\s*|[\\p{L}\\p{N}\\r\\n\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*[\\p{L}\\p{N}][\\p{L}\\p{N}\\r\\n\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*",
            "type" : "string"
          },
          "resolution" : {
            "description" : "Código de resolución",
            "example" : "FALSO_POSITIVO",
            "maxLength" : 30,
            "minLength" : 0,
            "pattern" : "[A-Z0-9][A-Z0-9_-]*",
            "type" : "string"
          }
        },
        "required" : [ "resolution" ],
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "description" : "Error de la API (RFC 9457)",
        "properties" : {
          "detail" : {
            "example" : "Onboarding request not found",
            "type" : "string"
          },
          "instance" : {
            "type" : "string"
          },
          "status" : {
            "example" : 404,
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "example" : "Not Found",
            "type" : "string"
          },
          "type" : {
            "example" : "about:blank",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RaiseAlertRequest" : {
        "properties" : {
          "accountId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "criticality" : {
            "enum" : [ "CRITICAL", "HIGH", "MEDIUM", "LOW" ],
            "type" : "string"
          },
          "evidence" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "reason" : {
            "maxLength" : 250,
            "minLength" : 0,
            "pattern" : "\\s*|[\\p{L}\\p{N}\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*[\\p{L}\\p{N}][\\p{L}\\p{N}\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*",
            "type" : "string"
          },
          "typeCode" : {
            "maxLength" : 40,
            "minLength" : 0,
            "pattern" : "[A-Z0-9][A-Z0-9_-]*",
            "type" : "string"
          }
        },
        "required" : [ "accountId", "reason", "typeCode" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token de POST /api/console/auth/login. Solo la consola lo usa; el flujo móvil es público.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API del flujo de onboarding móvil (apertura de cuenta), del score de riesgo y de la consola administrativa.\n\n- Rutas `/api/onboarding/**` y `/api/catalogs`: públicas (app móvil).\n- Rutas `/api/console/**`: requieren `Authorization: Bearer <token>` de `POST /api/console/auth/login`.\n- Errores: `application/problem+json` con `status`, `title` y `detail`.\n- Los campos que aparecen como opcionales en los esquemas de respuesta pueden venir `null`.",
    "title" : "Verificación Online API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/console/alert-types" : {
      "get" : {
        "operationId" : "alertTypes",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AlertTypeResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Tipos de alerta"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Tipos de alerta y su criticidad por defecto",
        "tags" : [ "Consola · Alertas" ]
      }
    },
    "/api/console/alerts" : {
      "get" : {
        "description" : "De la más crítica a la menos crítica y, dentro de cada criticidad, la más reciente primero. `account` busca por los últimos 4 dígitos; `from` y `to` son días (aaaa-mm-dd, hora de El Salvador), ambos inclusivos.",
        "operationId" : "alertList",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "UNASSIGNED", "ASSIGNED", "IN_REVIEW", "CLOSED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "assigneeId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "account",
          "required" : false,
          "schema" : {
            "maxLength" : 30,
            "minLength" : 0,
            "pattern" : "[0-9 *-]*",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AlertResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Alertas abiertas"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Filtro inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Alertas abiertas ordenadas por criticidad",
        "tags" : [ "Consola · Alertas" ]
      },
      "post" : {
        "description" : "Sin `criticality` se usa la del tipo de alerta.",
        "operationId" : "alertRaise",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RaiseAlertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description" : "Alerta creada sin responsable"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Datos inválidos o tipo de alerta desconocido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La cuenta no existe"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Levanta una alerta sobre una cuenta",
        "tags" : [ "Consola · Alertas" ]
      }
    },
    "/api/console/alerts/{alertId}" : {
      "get" : {
        "description" : "Cuenta, tipo, evidencia, responsable y fechas del flujo.",
        "operationId" : "alertGet",
        "parameters" : [ {
          "in" : "path",
          "name" : "alertId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description" : "Alerta"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La alerta no existe"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Detalle de una alerta",
        "tags" : [ "Consola · Alertas" ]
      }
    },
    "/api/console/alerts/{alertId}/close" : {
      "post" : {
        "description" : "Puede cerrarla quien la tomó, la líder de Conozca a su Cliente o un administrador.",
        "operationId" : "alertClose",
        "parameters" : [ {
          "in" : "path",
          "name" : "alertId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CloseAlertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description" : "Alerta cerrada"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta la resolución"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "El usuario no puede cerrar esta alerta"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La alerta no existe"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La alerta no está asignada ni en revisión"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Cierra una alerta con su resolución",
        "tags" : [ "Consola · Alertas" ]
      }
    },
    "/api/console/alerts/{alertId}/history" : {
      "get" : {
        "operationId" : "alertHistory",
        "parameters" : [ {
          "in" : "path",
          "name" : "alertId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AlertHistoryResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Cambios de estado, del más antiguo al más reciente"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La alerta no existe"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Historial de una alerta",
        "tags" : [ "Consola · Alertas" ]
      }
    },
    "/api/console/alerts/{alertId}/review" : {
      "post" : {
        "description" : "Solo quien la tomó.",
        "operationId" : "alertStartReview",
        "parameters" : [ {
          "in" : "path",
          "name" : "alertId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AlertCommentRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description" : "La alerta quedó en revisión"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "El usuario no es el responsable de la alerta"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La alerta no existe"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La alerta no está asignada"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Empieza a revisar una alerta",
        "tags" : [ "Consola · Alertas" ]
      }
    },
    "/api/console/alerts/{alertId}/take" : {
      "post" : {
        "description" : "Se la asigna al usuario de la sesión y queda en el historial de la alerta. Solo analistas de fraude y administradores.",
        "operationId" : "alertTake",
        "parameters" : [ {
          "in" : "path",
          "name" : "alertId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AlertResponse"
                }
              }
            },
            "description" : "La alerta quedó asignada al usuario de la sesión"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "El rol del usuario no puede tomar alertas"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La alerta no existe"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La alerta ya fue tomada por alguien más o está cerrada"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Tomar una alerta sin dueño",
        "tags" : [ "Consola · Alertas" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Dev",
    "url" : "https://api.dev.identidad.alambritos.online"
  }, {
    "description" : "QA",
    "url" : "https://api.qa.identidad.alambritos.online"
  }, {
    "description" : "Local",
    "url" : "http://localhost:8080"
  } ],
...
}